feat: allow optional uuid template on job name (#209)
if a uuid is defined in the template view, it will append it to the job name. If it is not defined, it will keep the same name it always has used `razeedeploy-job`.

This will be useful, so users dont have to delete the old job if they want to run the job again.